+++ packages/generator/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.1.6
+- Don't check working tree when looking for changed packages
 - Register plugins in changed script.
 
 ## 0.1.5