+++ packages/generator/CHANGELOG.md
@@ -2,6 +2,9 @@
 - Added confirm changes prompt.
 - List uncommited files in working tree check.
 
+### Breaking changes
+- Default branch config was changed from `master` to `main`.
+
 ## 0.1.6
 - Don't check working tree when looking for changed packages
 - Register plugins in changed script.